fixes #73 - gatherer_url and image_url added back to tutor.card; using either name or id based on the search
uses the information (name or id) that searched up the card to construct the URL. As an implementation detail, variations in capitalization or punctuation can effect the resulting URL (although a URL will be returned iff the search string did successfully yield the card). This implementation may not work well with redirects, but those are currently not being followed.